cli: refactor refresh to the new approach
