+++ uaclient/cli/tests/test_cli_config.py
@@ -23,7 +23,7 @@ Available Commands:
 
 @mock.patch("uaclient.cli.LOG.error")
 @mock.patch("uaclient.log.setup_cli_logging")
-@mock.patch(M_PATH + "contract.get_available_resources")
+@mock.patch("uaclient.contract.get_available_resources")
 class TestMainConfig:
     @pytest.mark.parametrize("additional_params", ([], ["--help"]))
     def test_config_help(
