+++ uaclient/cli/tests/test_cli_refresh.py
@@ -2,47 +2,13 @@ import mock
 import pytest
 
 from uaclient import exceptions, lock, messages
-from uaclient.cli import action_refresh, main
+from uaclient.cli.refresh import refresh_command
 from uaclient.files.notices import Notice
 
-HELP_OUTPUT = """\
-usage: pro refresh [contract|config|messages] [flags]
-
-Refresh three distinct Ubuntu Pro related artifacts in the system:
-
-* contract: Update contract details from the server.
-* config:   Reload the config file.
-* messages: Update APT and MOTD messages related to UA.
-
-You can individually target any of the three specific actions,
-by passing the target name to the command.  If no `target`
-is specified, all targets are refreshed.
-
-positional arguments:
-  {contract,config,messages}
-                        Target to refresh.
-
-Flags:
-  -h, --help            show this help message and exit
-"""
+M_PATH = "uaclient.cli.refresh."
 
 
 class TestActionRefresh:
-    @mock.patch("uaclient.log.setup_cli_logging")
-    @mock.patch("uaclient.cli.contract.get_available_resources")
-    def test_refresh_help(
-        self, _m_resources, _m_setup_logging, capsys, FakeConfig
-    ):
-        with pytest.raises(SystemExit):
-            with mock.patch("sys.argv", ["/usr/bin/ua", "refresh", "--help"]):
-                with mock.patch(
-                    "uaclient.config.UAConfig",
-                    return_value=FakeConfig(),
-                ):
-                    main()
-        out, _err = capsys.readouterr()
-        assert HELP_OUTPUT in out
-
     @mock.patch("uaclient.util.we_are_currently_root", return_value=False)
     def test_non_root_users_are_rejected(
         self, we_are_currently_root, FakeConfig, fake_machine_token_file
@@ -52,7 +18,7 @@ class TestActionRefresh:
         cfg = FakeConfig()
         fake_machine_token_file.attached = True
         with pytest.raises(exceptions.NonRootUserError):
-            action_refresh(mock.MagicMock(), cfg=cfg)
+            refresh_command.action(mock.MagicMock(), cfg=cfg)
 
     @pytest.mark.parametrize(
         "target, expect_unattached_error",
@@ -75,9 +41,11 @@ class TestActionRefresh:
         with mock.patch.object(lock, "lock_data_file"):
             if expect_unattached_error:
                 with pytest.raises(exceptions.UnattachedError):
-                    action_refresh(mock.MagicMock(target=target), cfg=cfg)
+                    refresh_command.action(
+                        mock.MagicMock(target=target), cfg=cfg
+                    )
             else:
-                action_refresh(mock.MagicMock(target=target), cfg=cfg)
+                refresh_command.action(mock.MagicMock(target=target), cfg=cfg)
 
     @mock.patch("uaclient.lock.check_lock_info")
     @mock.patch("time.sleep")
@@ -95,7 +63,7 @@ class TestActionRefresh:
         fake_machine_token_file.attached = True
         m_check_lock_info.return_value = (123, "pro disable")
         with pytest.raises(exceptions.LockHeldError) as err:
-            action_refresh(mock.MagicMock(), cfg=cfg)
+            refresh_command.action(mock.MagicMock(), cfg=cfg)
         assert 12 == m_check_lock_info.call_count
         expected_msg = messages.E_LOCK_HELD_ERROR.format(
             lock_request="pro refresh", lock_holder="pro disable", pid=123
@@ -125,7 +93,9 @@ class TestActionRefresh:
 
         with pytest.raises(exceptions.UbuntuProError) as excinfo:
             with mock.patch.object(lock, "lock_data_file"):
-                action_refresh(mock.MagicMock(target="contract"), cfg=cfg)
+                refresh_command.action(
+                    mock.MagicMock(target="contract"), cfg=cfg
+                )
 
         assert messages.E_REFRESH_CONTRACT_FAILURE.msg == excinfo.value.msg
         assert [
@@ -150,7 +120,9 @@ class TestActionRefresh:
         cfg = FakeConfig()
         fake_machine_token_file.attached = True
         with mock.patch.object(lock, "lock_data_file"):
-            ret = action_refresh(mock.MagicMock(target="contract"), cfg=cfg)
+            ret = refresh_command.action(
+                mock.MagicMock(target="contract"), cfg=cfg
+            )
 
         assert 0 == ret
         assert messages.REFRESH_CONTRACT_SUCCESS in capsys.readouterr()[0]
@@ -160,7 +132,7 @@ class TestActionRefresh:
         ] == m_remove_notice.call_args_list
 
     @mock.patch("uaclient.lock.check_lock_info", return_value=(-1, ""))
-    @mock.patch("uaclient.cli.update_motd_messages")
+    @mock.patch(M_PATH + "update_motd_messages")
     def test_refresh_messages_error(
         self, m_update_motd, _m_check_lock_info, FakeConfig
     ):
@@ -169,7 +141,7 @@ class TestActionRefresh:
 
         with pytest.raises(exceptions.UbuntuProError) as excinfo:
             with mock.patch.object(lock, "lock_data_file"):
-                action_refresh(
+                refresh_command.action(
                     mock.MagicMock(target="messages"), cfg=FakeConfig()
                 )
 
@@ -180,7 +152,7 @@ class TestActionRefresh:
     @mock.patch("uaclient.timer.update_messaging.exists", return_value=True)
     @mock.patch("uaclient.timer.update_messaging.LOG.exception")
     @mock.patch("uaclient.system.subp")
-    @mock.patch("uaclient.cli.update_motd_messages")
+    @mock.patch(M_PATH + "update_motd_messages")
     def test_refresh_messages_doesnt_fail_if_update_notifier_does(
         self,
         m_update_motd,
@@ -196,7 +168,7 @@ class TestActionRefresh:
         m_subp.side_effect = [subp_exc, ""]
 
         with mock.patch.object(lock, "lock_data_file"):
-            ret = action_refresh(
+            ret = refresh_command.action(
                 mock.MagicMock(target="messages"), cfg=FakeConfig()
             )
 
@@ -207,8 +179,8 @@ class TestActionRefresh:
 
     @mock.patch("uaclient.lock.check_lock_info", return_value=(-1, ""))
     @mock.patch("uaclient.apt_news.update_apt_news")
-    @mock.patch("uaclient.cli.refresh_motd")
-    @mock.patch("uaclient.cli.update_motd_messages")
+    @mock.patch(M_PATH + "refresh_motd")
+    @mock.patch(M_PATH + "update_motd_messages")
     def test_refresh_messages_happy_path(
         self,
         m_update_motd,
@@ -221,7 +193,9 @@ class TestActionRefresh:
         """On success from request_updates_contract root user can refresh."""
         cfg = FakeConfig()
         with mock.patch.object(lock, "lock_data_file"):
-            ret = action_refresh(mock.MagicMock(target="messages"), cfg=cfg)
+            ret = refresh_command.action(
+                mock.MagicMock(target="messages"), cfg=cfg
+            )
 
         assert 0 == ret
         assert messages.REFRESH_MESSAGES_SUCCESS in capsys.readouterr()[0]
@@ -251,7 +225,9 @@ class TestActionRefresh:
 
         with pytest.raises(exceptions.UbuntuProError) as excinfo:
             with mock.patch.object(lock, "lock_data_file"):
-                action_refresh(mock.MagicMock(target="config"), cfg=cfg)
+                refresh_command.action(
+                    mock.MagicMock(target="config"), cfg=cfg
+                )
 
         assert messages.E_REFRESH_CONFIG_FAILURE.msg == excinfo.value.msg
 
@@ -270,7 +246,9 @@ class TestActionRefresh:
         cfg = FakeConfig()
         fake_machine_token_file.attached = True
         with mock.patch.object(lock, "lock_data_file"):
-            ret = action_refresh(mock.MagicMock(target="config"), cfg=cfg)
+            ret = refresh_command.action(
+                mock.MagicMock(target="config"), cfg=cfg
+            )
 
         assert 0 == ret
         assert messages.REFRESH_CONFIG_SUCCESS in capsys.readouterr()[0]
@@ -278,8 +256,8 @@ class TestActionRefresh:
 
     @mock.patch("uaclient.lock.check_lock_info", return_value=(-1, ""))
     @mock.patch("uaclient.apt_news.update_apt_news")
-    @mock.patch("uaclient.cli.refresh_motd")
-    @mock.patch("uaclient.cli.update_motd_messages")
+    @mock.patch(M_PATH + "refresh_motd")
+    @mock.patch(M_PATH + "update_motd_messages")
     @mock.patch("uaclient.contract.refresh")
     @mock.patch("uaclient.config.UAConfig.process_config")
     @mock.patch("uaclient.files.notices.NoticesManager.remove")
@@ -301,7 +279,7 @@ class TestActionRefresh:
         cfg = FakeConfig()
         fake_machine_token_file.attached = True
         with mock.patch.object(lock, "lock_data_file"):
-            ret = action_refresh(mock.MagicMock(target=None), cfg=cfg)
+            ret = refresh_command.action(mock.MagicMock(target=None), cfg=cfg)
 
         out, err = capsys.readouterr()
